+++ source/js/vendors.js
@@ -1,0 +1,3 @@
+//= include vendors/skip-link-focus-fix.js
+//= include vendors/fitvids.js
+//= include vendors/slick.js